pyproject.toml
[tool.poetry]
name = "maddie"
version = "0.1.0"
description = "Masks Automated Discord Dice Interpreter & Explainer"
authors = ["Harkano", "also, others"]
license = "GPL-3.0"

[tool.poetry.dependencies]
# Ideally, we should test with various versions of Python
# and see which ones Maddie works with. This is just the one
# I have on my laptop.
python = "^3.8"
"discord.py" = "^1.5.1"
python-dotenv = "^0.15.0"
boto3 = "^1.16.7"
tssplit = "^1.0.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
